test: clean up api db mocks and server status (#79)
* test: clean up api db mocks and server status

* chore: remove todo
