+++ src/server-version.ts
@@ -7,6 +7,13 @@ interface ServerVersion {
 }
 
 function getServerVersion(): ServerVersion {
+  if (process.env.NODE_ENV === 'test') {
+    return {
+      branch: 'test',
+      commit: '123456',
+      tag: 'v0.0.1',
+    };
+  }
   const [branch, commit, tag] = readFileSync('.git-info', 'utf-8').split('\n');
   return { branch, commit, tag };
 }
