+++ tests/blockchain-smart-contract-monitor.test.ts
@@ -1,15 +1,14 @@
 import { bufferCV, cvToHex, listCV, stringUtf8CV, tupleCV, uintCV } from '@stacks/transactions';
-import * as postgres from 'postgres';
 import { ENV } from '../src/env';
 import {
   BlockchainDbContractLog,
   BlockchainDbSmartContract,
-  PgBlockchainApiStore,
 } from '../src/pg/blockchain-api/pg-blockchain-api-store';
 import { PgStore } from '../src/pg/pg-store';
 import { DbSipNumber, DbSmartContractInsert, DbTokenType } from '../src/pg/types';
 import { BlockchainSmartContractMonitor } from '../src/token-processor/blockchain-api/blockchain-smart-contract-monitor';
 import { cycleMigrations } from '../src/pg/migrations';
+import { MockPgBlockchainApiStore } from './helpers';
 
 const NftAbi = {
   maps: [],
@@ -162,28 +161,6 @@ const NftAbi = {
   non_fungible_tokens: [{ name: 'friedger-pool', type: 'uint128' }],
 };
 
-class MockPgBlockchainApiStore extends PgBlockchainApiStore {
-  public smartContract?: BlockchainDbSmartContract;
-  public contractLog?: BlockchainDbContractLog;
-
-  constructor(smartContract?: BlockchainDbSmartContract, contractLog?: BlockchainDbContractLog) {
-    super(postgres());
-    this.smartContract = smartContract;
-    this.contractLog = contractLog;
-  }
-
-  getSmartContract(args: { contractId: string }): Promise<BlockchainDbSmartContract | undefined> {
-    return Promise.resolve(this.smartContract);
-  }
-
-  getSmartContractLog(args: {
-    txId: string;
-    eventIndex: number;
-  }): Promise<BlockchainDbContractLog | undefined> {
-    return Promise.resolve(this.contractLog);
-  }
-}
-
 class TestBlockchainMonitor extends BlockchainSmartContractMonitor {
   public async testHandleMessage(message: string) {
     return this.handleMessage(message);
@@ -210,7 +187,8 @@ describe('BlockchainSmartContractMonitor', () => {
       block_height: 1,
       abi: NftAbi,
     };
-    const apiDb = new MockPgBlockchainApiStore(contract);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.smartContract = contract;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
     await monitor.testHandleMessage(
       JSON.stringify({
@@ -237,7 +215,8 @@ describe('BlockchainSmartContractMonitor', () => {
       block_height: 1,
       abi: { maps: [], functions: [], variables: [], fungible_tokens: [], non_fungible_tokens: [] },
     };
-    const apiDb = new MockPgBlockchainApiStore(contract);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.smartContract = contract;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
     await monitor.testHandleMessage(
       JSON.stringify({
@@ -289,7 +268,8 @@ describe('BlockchainSmartContractMonitor', () => {
         })
       ),
     };
-    const apiDb = new MockPgBlockchainApiStore(undefined, event);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.contractLog = event;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
 
     await monitor.testHandleMessage(
@@ -341,7 +321,8 @@ describe('BlockchainSmartContractMonitor', () => {
         })
       ),
     };
-    const apiDb = new MockPgBlockchainApiStore(undefined, event);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.contractLog = event;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
 
     await monitor.testHandleMessage(
@@ -365,7 +346,8 @@ describe('BlockchainSmartContractMonitor', () => {
       sender_address: 'SP1K1A1PMGW2ZJCNF46NWZWHG8TS1D23EGH1KNK60',
       value: cvToHex(stringUtf8CV('test')),
     };
-    const apiDb = new MockPgBlockchainApiStore(undefined, event);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.contractLog = event;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
     await monitor.testHandleMessage(
       JSON.stringify({
@@ -421,7 +403,8 @@ describe('BlockchainSmartContractMonitor', () => {
         })
       ),
     };
-    const apiDb = new MockPgBlockchainApiStore(undefined, event);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.contractLog = event;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
 
     await monitor.testHandleMessage(
@@ -473,7 +456,8 @@ describe('BlockchainSmartContractMonitor', () => {
         })
       ),
     };
-    const apiDb = new MockPgBlockchainApiStore(undefined, event);
+    const apiDb = new MockPgBlockchainApiStore();
+    apiDb.contractLog = event;
     const monitor = new TestBlockchainMonitor({ db, apiDb });
 
     await monitor.testHandleMessage(
